% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_sizes.R
\name{file_size_shiny}
\alias{file_size_shiny}
\title{File size density plot server function}
\usage{
file_size_shiny(input, output, session, file_sizes = sizes_log)
}
\description{
File size density plot server function
}

