% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_sizes.R
\name{file_sizes_db}
\alias{file_sizes_db}
\title{Grouped file size query}
\usage{
file_sizes_db()
}
\description{
Memoised because it will be called repeatedly from \code{\link{file_sizes}}
This value is cached in this repository as file_sizes_log
}

