% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_sizes.R
\name{range_size_percentage}
\alias{range_size_percentage}
\title{Get the percentage ratio of a file size range}
\usage{
range_size_percentage(file_sizes, range)
}
\arguments{
\item{file_sizes}{All file sizes}

\item{range}{Numeric vector of two, giving a range for file sizes.}
}
\value{
The percentage of file sizes that are between \code{range}.
}
\description{
Get the percentage ratio of a file size range
}

